remove option auto-compile-toggle-recursively
For now when the user selects a directory always process
subdirectories recursively.

There might be situations when this is not desirable but normally
toggling recursively is prefered and this should therefor not be
controlled using a permanent setting.

Toggling the state of files in a directory but not it's subdirectories
might come back to toggle-auto-compile by means of prefix arguments.
Though at the moment I am not sure whether that would be all that
useful; it might just make the doc-string even more verbose.