auto-compile-handle-compile-error: don't start compiling
Previously when this function was called due to an unbalanced
expression, `auto-compile-pretend-byte-compiled' was unconditionally set
to `t'.  Fixing the expression and saving again, then caused the file to
be byte-compiled.  Now we only set `auto-compile-pretend-byte-compiled'
on failure when the new START argument is non-nil, or when the
byte-compile destination already exists.

Fixes #12.