Fix use of autoload-modified-buffers
This is supposed to list the files where we put the definitions not
the files from where we collect the cookies.  Once we stop misusing
this variable we can also stop advising `autoload-save-buffers',
which we only had to undo the effects of this mistake.

We continue to let-bind `autoload-modified-buffers' but now we use
the empty list to prevent any unrelated buffers from being saved,
which was the intention before as well.