Drop dependency on packed
As a consequence `toggle-auto-compile' now compiles all emacs-lisp
source files, not just libraries that provide the expected features.

A new option `auto-compile-predicate-function' allows restoring the
old behavior, by using `elx-library-p' from the `elx' package, which
makes it necessary to install that package.

This makes it possible to deprecate `packed'.  Most of the functions
that I actually use were moved to `elx'.  A few functions were moved
or copied to this package.  Some unused functionality was dropped.

One of the goals behind these changes, was to make this package not
depend on a package that has been bitrotting for a while already.