Fix compilation of init file during startup
`startup--load-user-init-file' binds `user-init-file' to t, and then
calls `load', which (due to that special value) sets that variable to
the file it loads.

`auto-compile-on-load' calls `byte-compile-file', which in this case
causes other files to be loaded.  We have to suspend the special
`user-init-file' value, or the first of them thinks it is loaded the
init file and sets thus sets `user-init-file' to the file it loaded.

Closes #33.