Only use global-font-lock-mode-enable-in-buffer if available
Since [1: 9506b9392e8], Emacs 30's `define-globalized-minor-mode'
defines and relies on this function.  The issue that made that
change necessary presumably doesn't exist on older Emacs releases,
or at least I cannot remember, running into the issue before, which
lead to [2: 4a951af], which fixed the issue for Emacs 30, but caused
a void-function error on older releases.

Closes #36.

1: 2024-04-08 9506b9392e8c3548b769d9951a1c9d4abad9ee21
   (define-globalized-minor-mode): Fix bug#69431

2: 2024-06-08 4a951af6b186ce86d324460d6629bc0ed391f482
   auto-compile--byte-compile-file: Restore font-lock activation